refactor(libpod): extract error.rs tests into a separate file
The error.rs file grew past 500 code lines after the #1357 Field
variant + Display arm + Field error source arm + new tests. Move
the three test modules out of error.rs to error/tests.rs so the
production file is under the line cap. The test file imports
`super::` items through `crate::libpod::error::` (Rust's test_modules
inside a non-test module cannot import from the parent module the same
way; the path is explicit).

Pure mechanical move. No semantic change.
